没什么特别意义的一个版本

+++ Common/ReturnCode.cs
@@ -1,6 +1,6 @@
 ﻿namespace Common
 {
-    public enum ReturnCode:byte
+    public enum ReturnCode:ushort
     {
         Success,
         Failed
